refactor: Players 클래스의 비즈니스 로직에 관한 코드를 분리하여 PlayersService 클래스 생성
